+++ app/src/components/Cinema/Cinema.tsx
@@ -3,7 +3,6 @@
 // SPDX-License-Identifier: EUPL-1.2
 import { styled, Container as MuiContainer } from '@mui/material';
 import { useSnackbarFacade } from '@opentalk/common';
-import React from 'react';
 import { FullScreen as ReactFullScreen } from 'react-full-screen';
 
 import LayoutOptions from '../../enums/LayoutOptions';