+++ app/src/components/TimerTab/fragments/TimerOverview.tsx
@@ -2,13 +2,13 @@
 //
 // SPDX-License-Identifier: EUPL-1.2
 import { Button, Stack, styled, Typography } from '@mui/material';
-import React from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { stopTimer } from '../../../api/types/outgoing/timer';
+import { stopTimer, TimerStyle } from '../../../api/types/outgoing/timer';
 import { useAppDispatch, useAppSelector } from '../../../hooks';
 import { selectTimerId } from '../../../store/slices/timerSlice';
-import TimerCounter from './TimerCounter';
+import TimerCoffeeBreakCounter from './TimerCoffeeBreakCounter';
+import TimerNormalCounter from './TimerNormalCounter';
 import UserList from './UserList';
 
 const Container = styled('div')(({ theme }) => ({
@@ -18,15 +18,21 @@ const Container = styled('div')(({ theme }) => ({
   gap: theme.spacing(2),
   overflow: 'auto',
   gridTemplateColumns: '1fr',
-  gridTemplateRows: 'repeat(2, auto) 1fr auto',
+  gridTemplateRows: 'auto 1fr auto',
   justifyContent: 'center',
+}));
+
+const BoxNormal = styled('div')(({ theme }) => ({
+  display: 'flex',
+  flexDirection: 'column',
+  gap: theme.spacing(2),
 }));
 
 const TabTitle = styled(Typography)(({ theme }) => ({
   padding: theme.spacing(1, 2),
 }));
 
-const TimerOverview = () => {
+const TimerOverview = ({ timerStyle }: { timerStyle: TimerStyle }) => {
   const { t } = useTranslation();
   const dispatch = useAppDispatch();
   const timerId = useAppSelector(selectTimerId);
@@ -39,9 +45,23 @@ const TimerOverview = () => {
 
   return (
     <Container>
-      <TabTitle variant={'h2'}>{t('timer-header-title')}</TabTitle>
-      <TimerCounter />
-      <UserList />
+      {timerStyle === TimerStyle.CoffeeBreak && (
+        <>
+          <Typography variant={'h2'}>{t('coffee-break-header-title')}</Typography>
+          <TimerCoffeeBreakCounter />
+        </>
+      )}
+
+      {timerStyle === TimerStyle.Normal && (
+        <>
+          <TabTitle variant={'h2'}>{t('timer-header-title')}</TabTitle>
+          <BoxNormal>
+            <TimerNormalCounter />
+            <UserList />
+          </BoxNormal>
+        </>
+      )}
+
       <Stack justifySelf={'center'}>
         <Button color="secondary" onClick={handleStop}>
           {t('timer-overview-button-stop')}