+++ app/src/store/slices/timerSlice.ts
@@ -5,7 +5,14 @@ import { ParticipantId, Timestamp } from '@opentalk/common';
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
 import { RootState } from '../';
-import { StartTimer, ReadyToContinue } from '../../api/types/incoming/timer';
+import { StartTimer, ReadyToContinue, TimerStopKind } from '../../api/types/incoming/timer';
+import { TimerState } from './../../api/types/incoming/control';
+import { TimerKind, TimerStyle } from './../../api/types/outgoing/timer';
+
+interface RealTime {
+  format: string;
+  duration: Duration;
+}
 
 interface State {
   startedAt?: Timestamp;
@@ -15,6 +22,12 @@ interface State {
   readyCheckEnabled: boolean;
   title?: string;
   running: boolean;
+  message?: 'started' | 'stopped';
+  kind?: TimerKind;
+  style?: TimerStyle;
+  realTime?: RealTime;
+  initialTime?: RealTime;
+  kindStopTimer?: TimerStopKind;
 }
 
 const initialState = {
@@ -25,12 +38,19 @@ const initialState = {
   readyCheckEnabled: false,
   title: undefined,
   running: false,
+  kind: undefined,
+  style: undefined,
+  message: undefined,
+  realTime: undefined,
+  initialTime: undefined,
+  kindStopTimer: undefined,
 };
 
 export const timerSlice = createSlice({
   name: 'timer',
   initialState: initialState as State,
   reducers: {
+    setInitialStateTimer: () => initialState,
     startedTimer: (
       state,
       { payload: { payload, timestamp } }: PayloadAction<{ payload: StartTimer; timestamp: Timestamp }>
@@ -41,10 +61,24 @@ export const timerSlice = createSlice({
       state.readyCheckEnabled = payload.readyCheckEnabled;
       state.title = payload.title;
       state.running = true;
+      state.message = payload.message;
+      state.kind = payload.kind;
+      state.style = payload.style;
     },
-    stoppedTimer: (state) => {
+    joinedTimer: (state, { payload }: PayloadAction<TimerState>) => {
+      state.endsAt = payload.endsAt;
+      state.kind = payload.kind;
+      state.readyCheckEnabled = payload.readyCheckEnabled;
+      state.startedAt = payload.startedAt;
+      state.style = payload.style;
+      state.timerId = payload.timerId;
+      state.message = 'started';
+      state.running = true;
+    },
+    stoppedTimer: (state, { payload }: { payload: { message: 'stopped'; kindStopTimer: TimerStopKind } }) => {
       state.running = false;
-      state.participantsReady = [];
+      state.message = payload.message;
+      state.kindStopTimer = payload.kindStopTimer;
     },
     updateParticpantsReady: (state, { payload }: PayloadAction<ReadyToContinue>) => {
       if (payload.status === true && !state.participantsReady.includes(payload.participantId)) {
@@ -55,10 +89,24 @@ export const timerSlice = createSlice({
         state.participantsReady = state.participantsReady.filter((item) => item !== payload.participantId);
       }
     },
+    setRealTime: (state, action) => {
+      state.realTime = action.payload as RealTime;
+    },
+    setInitialTime: (state, action) => {
+      state.initialTime = action.payload as RealTime;
+    },
   },
 });
 
-export const { startedTimer, stoppedTimer, updateParticpantsReady } = timerSlice.actions;
+export const {
+  startedTimer,
+  stoppedTimer,
+  updateParticpantsReady,
+  setRealTime,
+  setInitialTime,
+  setInitialStateTimer,
+  joinedTimer,
+} = timerSlice.actions;
 
 export const actions = timerSlice.actions;
 
@@ -69,5 +117,11 @@ export const selectParticipantsReady = (state: RootState) => state.timer.partici
 export const selectReadyCheckEnabled = (state: RootState) => state.timer.readyCheckEnabled;
 export const selectTimerTitle = (state: RootState) => state.timer.title;
 export const selectTimerRunning = (state: RootState) => state.timer.running;
+export const selectTimerKind = (state: RootState) => state.timer.kind;
+export const selectTimerStyle = (state: RootState) => state.timer.style;
+export const selectInitialTime = (state: RootState) => state.timer.initialTime;
+export const selectRealTime = (state: RootState) => state.timer.realTime;
+export const selectTimerMessage = (state: RootState) => state.timer.message;
+export const selectTimerStopKind = (state: RootState) => state.timer.kindStopTimer;
 
 export default timerSlice.reducer;